move models to module folder
