+++ Cargo.toml
@@ -12,7 +12,7 @@ base64 = "0.13.0"
 chrono = "0.4"
 clap = { version = "3.0.13", features = ["derive"] }
 futures = "0.3" # for our async / await blocks
-reqwest = { version = "0.11", features = ["json", "cookies", "blocking"] } # reqwest with JSON parsing support
+reqwest = { version = "0.11", features = ["json", "cookies"] } # reqwest with JSON parsing support
 serde = "1.0.136"
 serde_json = "1.0"
 tabled = "0.5"
