+++ src/model/bundle.rs
@@ -1,7 +1,7 @@
 use serde::Deserialize;
 use tabled::Tabled;
 
-use crate::models::null_to_default;
+use crate::model::null_to_default;
 
 #[allow(dead_code)]
 #[derive(Debug, Tabled, Deserialize)]
