+++ src/model/device.rs
@@ -1,65 +1,8 @@
-use std::{collections::HashMap, fmt};
-
-use serde::{Deserialize, Deserializer};
+use serde::Deserialize;
+use std::fmt;
 use tabled::Tabled;
 
-#[allow(dead_code)]
-#[derive(Tabled, Debug, Deserialize)]
-pub struct Appliance {
-    #[serde(deserialize_with = "null_to_default")]
-    pub add_time: i64,
-    #[serde(default)]
-    #[serde(deserialize_with = "null_to_default")]
-    pub advanced_analysis_capacity: i64,
-    #[serde(default)]
-    #[serde(deserialize_with = "null_to_default")]
-    pub analysis_levels_managed: bool,
-    #[serde(default)]
-    #[serde(deserialize_with = "null_to_default")]
-    pub total_capacity: i64,
-    #[serde(deserialize_with = "null_to_default")]
-    pub fingerprint: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub connection_type: String,
-    pub data_access: bool,
-    #[serde(default)]
-    #[serde(deserialize_with = "null_to_default")]
-    pub display_name: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub firmware_version: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub hostname: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub id: i64,
-    #[serde(deserialize_with = "null_to_default")]
-    pub license_status: String,
-    pub managed_by_local: bool,
-    pub manages_local: bool,
-    #[serde(deserialize_with = "null_to_default")]
-    pub nickname: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub platform: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub status_message: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub sync_time: i64,
-    #[serde(deserialize_with = "null_to_default")]
-    pub uuid: String,
-}
-
-
-#[allow(dead_code)]
-#[derive(Debug, Tabled, Deserialize)]
-pub struct Customization {
-    pub auto: bool,
-    #[serde(deserialize_with = "null_to_default")]
-    pub create_time: i64,
-    #[serde(deserialize_with = "null_to_default")]
-    pub id: i64,
-    #[serde(deserialize_with = "null_to_default")]
-    pub name: String,
-    pub recovered: bool,
-}
+use crate::model::null_to_default;
 
 #[allow(dead_code)]
 #[derive(Tabled, Debug, Deserialize)]
@@ -147,120 +90,6 @@ pub struct Device {
     pub vlanid: i64,
     #[serde(deserialize_with = "null_to_default")]
     pub vpc_id: String,
-}
-
-#[allow(dead_code)]
-#[derive(Tabled, Debug, Deserialize)]
-pub struct ExtraHop {
-    #[serde(deserialize_with = "null_to_default")]
-    pub display_host: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub external_hostname: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub hostname: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub mgmt_ipaddr: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub platform: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub version: String,
-}
-
-#[allow(dead_code)]
-#[derive(Tabled, Debug, Deserialize)]
-pub struct License {
-    #[serde(deserialize_with = "null_to_default")]
-    pub dossier: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub edition: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub expires_at: i64,
-    // pub modules: LicenseModules,
-    // pub options: LicenseOptions,
-    pub expires_in: i64,
-    #[serde(deserialize_with = "null_to_default")]
-    pub platform: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub product_key: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub serial: String,
-}
-
-#[allow(dead_code)]
-#[derive(Debug, Deserialize)]
-pub struct LicenseModules {
-    #[serde(deserialize_with = "null_to_default")]
-    pub modules: HashMap<String, bool>,
-}
-
-#[allow(dead_code)]
-#[derive(Debug, Deserialize)]
-pub struct LicenseOptions {
-    #[serde(deserialize_with = "null_to_default")]
-    pub options: HashMap<String, bool>,
-}
-
-#[allow(dead_code)]
-#[derive(Tabled, Debug, Deserialize)]
-pub struct Network {
-    #[serde(deserialize_with = "null_to_default")]
-    pub appliance_uuid: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub id: i64,
-    #[serde(deserialize_with = "null_to_default")]
-    pub idle: bool,
-    #[serde(deserialize_with = "null_to_default")]
-    pub mod_time: i64,
-    #[serde(deserialize_with = "null_to_default")]
-    pub name: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub node_id: i64,
-}
-
-#[allow(dead_code)]
-#[derive(Debug, Deserialize)]
-pub struct RunningConfig {
-    // config: serde_json::Value,
-    pub json: serde_json::Value,
-}
-
-#[allow(dead_code)]
-#[derive(Tabled, Debug, Deserialize)]
-pub struct Tag {
-    #[serde(deserialize_with = "null_to_default")]
-    pub id: i64,
-    #[serde(deserialize_with = "null_to_default")]
-    pub mod_time: i64,
-    #[serde(deserialize_with = "null_to_default")]
-    pub name: String,
-}
-
-#[allow(dead_code)]
-#[derive(Tabled, Debug, Deserialize)]
-pub struct Vlan {
-    #[serde(deserialize_with = "null_to_default")]
-    pub description: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub id: i64,
-    #[serde(deserialize_with = "null_to_default")]
-    pub mod_time: i64,
-    #[serde(deserialize_with = "null_to_default")]
-    pub name: String,
-    #[serde(deserialize_with = "null_to_default")]
-    pub network_id: i64,
-    #[serde(deserialize_with = "null_to_default")]
-    pub node_id: i64,
-    #[serde(deserialize_with = "null_to_default")]
-    pub vlanid: i64,
-}
-
-fn null_to_default<'de, D, T>(deserializer: D) -> Result<T, D::Error>
-where
-    T: Default + Deserialize<'de>,
-    D: Deserializer<'de>,
-{
-    let opt = Option::deserialize(deserializer)?;
-    Ok(opt.unwrap_or_default())
 }
 
 impl fmt::Display for Device {
