+++ src/model/mod.rs
@@ -1,6 +1,15 @@
 use serde::{Deserialize, Deserializer};
 
+pub mod appliance;
 pub mod bundle;
+pub mod customization;
+pub mod device;
+pub mod extrahop;
+pub mod license;
+pub mod network;
+pub mod running_config;
+pub mod tag;
+pub mod vlan;
 
 pub fn null_to_default<'de, D, T>(deserializer: D) -> Result<T, D::Error>
 where
